# test: the regression test queries now run faster