--- a/test/expected/parameters.out
+++ b/test/expected/parameters.out
@@ -1,6 +1,6 @@
 -- shows common query parameters
 select *
-from jsonb_each(postgrest_openapi_spec('{test}')->'components'->'parameters')
+from jsonb_each(get_openapi_document('{test}')->'components'->'parameters')
 where key in ('select', 'order', 'limit', 'offset', 'on_conflict', 'columns', 'or', 'and', 'not.or', 'not.and');
      key     |                                                                                     value                                                                                     
 -------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -18,7 +18,7 @@ where key in ('select', 'order', 'limit', 'offset', 'on_conflict', 'columns', 'o
 
 -- shows common headers
 select key, jsonb_pretty(value)
-from jsonb_each(postgrest_openapi_spec('{test}')->'components'->'parameters')
+from jsonb_each(get_openapi_document('{test}')->'components'->'parameters')
 where key in ('preferGet', 'preferPost', 'preferPut', 'preferPatch', 'preferDelete', 'preferPostRpc', 'range');
       key      |                                 jsonb_pretty                                 
 ---------------+------------------------------------------------------------------------------
@@ -283,7 +283,7 @@ where key in ('preferGet', 'preferPost', 'preferPut', 'preferPatch', 'preferDele
 
 -- shows table columns as parameters
 select key, jsonb_pretty(value)
-from jsonb_each(postgrest_openapi_spec('{test}')->'components'->'parameters')
+from jsonb_each(get_openapi_document('{test}')->'components'->'parameters')
 where key like 'rowFilter.products.%';
               key               |        jsonb_pretty        
 --------------------------------+----------------------------
@@ -333,7 +333,7 @@ where key like 'rowFilter.products.%';
 
 -- shows view columns as parameters
 select key, jsonb_pretty(value)
-from jsonb_each(postgrest_openapi_spec('{test}')->'components'->'parameters')
+from jsonb_each(get_openapi_document('{test}')->'components'->'parameters')
 where key like 'rowFilter.big_products.%';
              key             |       jsonb_pretty       
 -----------------------------+--------------------------
@@ -369,7 +369,7 @@ where key like 'rowFilter.big_products.%';
 
 -- shows `IN` function arguments as RPC parameters
 select key, jsonb_pretty(value)
-from jsonb_each(postgrest_openapi_spec('{test}')->'components'->'parameters')
+from jsonb_each(get_openapi_document('{test}')->'components'->'parameters')
 where key like 'rpcParam.has_in_parameters.%';
              key              |              jsonb_pretty               
 ------------------------------+-----------------------------------------
@@ -413,7 +413,7 @@ where key like 'rpcParam.has_in_parameters.%';
 
 -- shows `INOUT` function arguments as RPC parameters
 select key, jsonb_pretty(value)
-from jsonb_each(postgrest_openapi_spec('{test}')->'components'->'parameters')
+from jsonb_each(get_openapi_document('{test}')->'components'->'parameters')
 where key like 'rpcParam.has_inout_parameters.%';
                key               |              jsonb_pretty               
 ---------------------------------+-----------------------------------------
@@ -457,7 +457,7 @@ where key like 'rpcParam.has_inout_parameters.%';
 
 -- shows `VARIADIC` function arguments as RPC parameters
 select key, jsonb_pretty(value)
-from jsonb_each(postgrest_openapi_spec('{test}')->'components'->'parameters')
+from jsonb_each(get_openapi_document('{test}')->'components'->'parameters')
 where key like 'rpcParam.has_variadic_parameter.%';
                 key                |           jsonb_pretty            
 -----------------------------------+-----------------------------------