--- a/test/expected/security.out
+++ b/test/expected/security.out
@@ -1,5 +1,5 @@
 -- shows the JWT security scheme
-select jsonb_pretty(postgrest_openapi_spec('{test}')->'components'->'securitySchemes'->'JWT');
+select jsonb_pretty(get_openapi_document('{test}')->'components'->'securitySchemes'->'JWT');
                                         jsonb_pretty                                        
 --------------------------------------------------------------------------------------------
  {                                                                                         +
@@ -11,7 +11,7 @@ select jsonb_pretty(postgrest_openapi_spec('{test}')->'components'->'securitySch
 (1 row)
 
 -- lists the JWT scheme as a security requirement object 
-select postgrest_openapi_spec('{test}')->'security' <@ '[{"JWT": []}]'::jsonb as has_jwt;
+select get_openapi_document('{test}')->'security' <@ '[{"JWT": []}]'::jsonb as has_jwt;
  has_jwt 
 ---------
  t