--- a/test/sql/reqbodies.sql
+++ b/test/sql/reqbodies.sql
@@ -1,24 +1,24 @@
 -- Tables
 
 -- defines an application/json request body
-select jsonb_pretty(postgrest_openapi_spec('{test}')->'components'->'requestBodies'->'products'->'content'->'application/json');
+select jsonb_pretty(get_openapi_document('{test}')->'components'->'requestBodies'->'products'->'content'->'application/json');
 
 -- defines an application/x-www-form-urlencoded request body
-select jsonb_pretty(postgrest_openapi_spec('{test}')->'components'->'requestBodies'->'products'->'content'->'application/x-www-form-urlencoded');
+select jsonb_pretty(get_openapi_document('{test}')->'components'->'requestBodies'->'products'->'content'->'application/x-www-form-urlencoded');
 
 -- defines a text/csv request body
-select jsonb_pretty(postgrest_openapi_spec('{test}')->'components'->'requestBodies'->'products'->'content'->'text/csv');
+select jsonb_pretty(get_openapi_document('{test}')->'components'->'requestBodies'->'products'->'content'->'text/csv');
 
 -- Views
 
 -- defines an application/json request body
-select jsonb_pretty(postgrest_openapi_spec('{test}')->'components'->'requestBodies'->'big_products'->'content'->'application/json');
+select jsonb_pretty(get_openapi_document('{test}')->'components'->'requestBodies'->'big_products'->'content'->'application/json');
 
 -- defines an application/x-www-form-urlencoded request body
-select jsonb_pretty(postgrest_openapi_spec('{test}')->'components'->'requestBodies'->'big_products'->'content'->'application/x-www-form-urlencoded');
+select jsonb_pretty(get_openapi_document('{test}')->'components'->'requestBodies'->'big_products'->'content'->'application/x-www-form-urlencoded');
 
 -- defines a text/csv request body
-select jsonb_pretty(postgrest_openapi_spec('{test}')->'components'->'requestBodies'->'big_products'->'content'->'text/csv');
+select jsonb_pretty(get_openapi_document('{test}')->'components'->'requestBodies'->'big_products'->'content'->'text/csv');
 
 -- does not define a request body for non auto-updatable views
-select postgrest_openapi_spec('{test}')->'components'->'requestBodies' ? 'non_auto_updatable' as value;
+select get_openapi_document('{test}')->'components'->'requestBodies' ? 'non_auto_updatable' as value;