… test/expected/info.out
… test/expected/paths.out
… test/expected/responses.out
… test/expected/schemas.out
… test/sql/paths.sql
… test/sql/schemas.sql
… test/sql/security.sql