It doesn't call the main function that generates the OpenAPI spec directly anymore. It saves the output in a table and the tests will query that table instead (via a function).

Test duration before this change:

```log
test info                         ... ok          504 ms
test main                         ... ok          221 ms
test parameters                   ... ok          692 ms
test paths                        ... ok         8585 ms
test reqbodies                    ... ok          794 ms
test responses                    ... ok         3781 ms
test schemas                      ... ok         4682 ms
test security                     ... ok          238 ms
test servers                      ... ok          435 ms
```

Test duration after this change:

```
test info                         ... ok            8 ms
test main                         ... ok          221 ms
test parameters                   ... ok            9 ms
test paths                        ... ok           27 ms
test reqbodies                    ... ok            8 ms
test responses                    ... ok           15 ms
test schemas                      ... ok           17 ms
test security                     ... ok            7 ms
test servers                      ... ok          431 ms
```